# bridge/utility/api.yaml
swagger: "2.0"
info:
  description: "This is an API for a wine-olive oil e-shop"
  version: "1.0.0"
  title: "e-shop"
  contact:
    email: "rs.lissa1997@gmail.com"
host: "localhost"
basePath: "/bridge/v1"
...
schemes:
  - "http"
paths:
...
  /list-products:
    get:
        tags:
          - "product"
        summary: "List all the products in the store"
        description: "List all the products in the store"
        operationId: "listProducts"
        produces:
          - "application/json"
        responses:
          "200":
            description: "successful operation"
            schema:
              $ref: "#/definitions/AllProducts"
          "404":
            description: "Products not found"
  /list-products-by-cart/{cartId}:
    get:
        tags:
          - "product"
        summary: " Given a cart id return a list of all the products in the cart"
        description: " Given a cart id return a list of all the products in the cart"
        operationId: "listProductsByCart"
        produces:
          - "application/json"
        parameters:
          - name: "cartId"
            in: "path"
            description: "Id of the cart"
            required: true
            type: "integer"
            format: "int64"
        responses:
          "200":
            description: "successful operation"
            schema:
              $ref: "#/definitions/AllProducts"
          "404":
            description: "Products not found"
...
  /list-countries:
    get:
      tags:
        - "country"
      summary: "List all the countries"
      operationId: "ListCountries"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/AllCountries"
        "404":
          description: "Countries not found"
...
  /list-subcountries/{countryCode}:
    get:
      tags:
        - "sub countries"
      summary: "List all the countries subdivisions of a specific country"
      operationId: "ListSubCountries"
      produces:
        - "application/json"
      parameters:
        - name: "countryCode"
          in: "path"
          description: "code of country to return"
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/AllSubCountries"
        "404":
          description: "Sub Countries not found"
definitions:
  Country:
    type: "object"
    properties:
      country_code:
        type: "string"
        example: "IT"
        maxLength: 2
        minLength: 2
      name:
        type: "string"
        example: "Italia"
  AllCountries:
    type: array
    items:
      $ref: '#/definitions/Country'
  CountrySubdivisions:
    type: "object"
    properties:
      id_subdivision:
        type: "integer"
        format: "int64"
      code:
        type: "string"
        example: "RG"
        maxLength: 2
        minLength: 2
      name:
        type: "string"
        example: "Ragusa"
  AllSubCountries:
    type: array
    items:
      $ref: '#/definitions/CountrySubdivisions'
...
  Product:
    type: "object"
    required:
      - "name"
      - "description"
      - "price"
      - "currencyCode"
      - "status"
    properties:
      id_product:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        example: "nero d'avola"
      description:
        type: "string"
        example: "è un vino molto buono"
      price:
        type: "integer"
        format: "int64"
        example: 15
      currencyCode:
        type: "string"
        maxLength: 3
        minLength: 3
        example: "EUR"
      status:
        type: "boolean"
      imageURL:
        type: "string"
        example: "https://i.ibb.co/L9p2D14/vino-rosso.jpg"
      stock:
        type: "integer"
        format: "int64"
        example: 30
      last_update:
        type: "string"
        format: "date-time"
...
  AllProducts:
    type: array
    items:
      $ref: '#/definitions/Product'
...